+++ flutter_navigation_generator/example/lib/main.dart
@@ -105,14 +105,11 @@ class _MyHomePageState extends State<MyHomePage> {
               child: const Text("Go to page 2 (with fade animation)"),
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator
-                  .goToExampleScreenWithRequiredArgument(
-                      data: [CustomModel('John', 25), CustomModel('Jeff', 27)]),
+              onPressed: () => mainNavigator.goToExampleScreenWithRequiredArgument(data: [CustomModel('John', 25), CustomModel('Jeff', 27)]),
               child: const Text("Go to ExampleScreenWithRequiredArgument"),
             ),
             ElevatedButton(
-              onPressed: () =>
-                  mainNavigator.customName(id: '1', name: 'John', age: 12),
+              onPressed: () => mainNavigator.customName(id: '1', name: 'John', age: 12),
               child: const Text("Go to RouteNameWithArguments"),
             ),
             ElevatedButton(
@@ -132,17 +129,14 @@ class _MyHomePageState extends State<MyHomePage> {
               child: const Text("Go to RouteNameWithArguments2"),
             ),
             ElevatedButton(
-              onPressed: () =>
-                  mainNavigator.showSheetRecursiveNavigationBottomSheet(),
+              onPressed: () => mainNavigator.showSheetRecursiveNavigationBottomSheet(),
               child: const Text("Show a bottom sheet with its own navigator"),
             ),
             ElevatedButton(
-              onPressed: () =>
-                  mainNavigator.showDialogExampleDialog(text: 'hi there'),
+              onPressed: () => mainNavigator.showDialogExampleDialog(text: 'hi there'),
               child: const Text("Show a full screen dialog"),
             ),
-            Text(
-                "Has a navigation blocked by a guard (not logged in): ${mainNavigator.canContinueNavigation()}"),
+            Text("Has a navigation blocked by a guard (not logged in): ${mainNavigator.canContinueNavigation()}"),
             if (mainNavigator.canContinueNavigation()) ...[
               ElevatedButton(
                 onPressed: () async {
@@ -365,9 +359,7 @@ class RecursiveNavigationBottomSheet extends StatelessWidget {
                 child: const Text("Go to second page"),
               ),
               ElevatedButton(
-                onPressed: () =>
-                    myNavigator.showSheetRecursiveNavigationBottomSheet(
-                        layers: layers + 1),
+                onPressed: () => myNavigator.showSheetRecursiveNavigationBottomSheet(layers: layers + 1),
                 child: const Text("Open another bottom sheet"),
               ),
             ],
@@ -472,8 +464,7 @@ class Error404 extends StatelessWidget {
               'We couldn\'t find this page, sorry :(',
             ),
             ElevatedButton(
-              onPressed: () =>
-                  mainNavigator.goToMyHomePage(title: 'returning from 404'),
+              onPressed: () => mainNavigator.goToMyHomePage(title: 'returning from 404'),
               child: const Text("go home"),
             ),
           ],
@@ -506,8 +497,7 @@ class ErrorNotLoggedIn extends StatelessWidget {
               'You are not logged in, sorry :(',
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.goToMyHomePage(
-                  title: 'returning from not logged in'),
+              onPressed: () => mainNavigator.goToMyHomePage(title: 'returning from not logged in'),
               child: const Text("go home"),
             ),
             ElevatedButton(
@@ -550,8 +540,7 @@ class LoggedInPage extends StatelessWidget {
               'You are logged in, yay :)',
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.goToMyHomePage(
-                  title: 'returning from logged in'),
+              onPressed: () => mainNavigator.goToMyHomePage(title: 'returning from logged in'),
               child: const Text("go home"),
             ),
           ],