+++ flutter_navigation_generator/lib/src/code_builder/navigator_builder.dart
@@ -5,6 +5,7 @@ import 'package:flutter_navigation_generator/src/code_builder/on_generate_route_
 import 'package:flutter_navigation_generator/src/code_builder/route_builder.dart';
 import 'package:flutter_navigation_generator/src/models/importable_type.dart';
 import 'package:flutter_navigation_generator/src/models/route_config.dart';
+import 'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart';
 
 class NavigatorBuilder {
   final Set<RouteConfig> routes;
@@ -13,6 +14,7 @@ class NavigatorBuilder {
   final ImportableType? pageType;
   final ImportableType? unknownRoute;
   final List<ImportableType> defaultGuards;
+  final IncludeQueryParametersType includeQueryParametersNavigatorConfig;
 
   NavigatorBuilder({
     required this.routes,
@@ -21,6 +23,7 @@ class NavigatorBuilder {
     required this.unknownRoute,
     required this.pageType,
     required this.defaultGuards,
+    required this.includeQueryParametersNavigatorConfig,
   });
 
   Field buildNavigatorKey() {
@@ -34,8 +37,7 @@ class NavigatorBuilder {
   }
 
   Spec generate() {
-    final hasGuards = routes.any((route) => route.guards?.isNotEmpty == true) ||
-        defaultGuards.isNotEmpty;
+    final hasGuards = routes.any((route) => route.guards?.isNotEmpty == true) || defaultGuards.isNotEmpty;
     return Mixin(
       (b) => b
         ..name = className
@@ -66,6 +68,7 @@ class NavigatorBuilder {
             routes: routes,
             pageType: pageType,
             targetFile: targetFile,
+            includeQueryParametersNavigatorConfig: includeQueryParametersNavigatorConfig,
           ).generate(),
         ),
     );