+++ flutter_navigation_generator/lib/src/models/route_config.dart
@@ -19,6 +19,7 @@ class RouteConfig {
   final List<ImportableType> parameters;
   final List<ImportableType>? guards;
   final Map<String, dynamic> defaultValues;
+  final IncludeQueryParametersType? includeQueryParameters;
 
   bool get routeNameContainsParameters => routeName.contains(':');
 
@@ -38,6 +39,7 @@ class RouteConfig {
     this.parameters = const [],
     this.guards,
     this.defaultValues = const {},
+    this.includeQueryParameters,
   });
 
   Map<String, dynamic> toMap() {
@@ -57,6 +59,7 @@ class RouteConfig {
       'parameters': parameters.map((x) => x.toMap()).toList(),
       'guards': guards?.map((x) => x.toMap()).toList(),
       'defaultValues': defaultValues,
+      'includeQueryParameters': includeQueryParameters?.index,
     };
   }
 
@@ -68,32 +71,20 @@ class RouteConfig {
       constructorName: map['constructorName'] ?? '',
       routeName: map['routeName'] ?? '',
       methodName: map['methodName'] ?? '',
-      routeNameIsDefinedByAnnotation:
-          map['routeNameIsDefinedByAnnotation'] ?? '',
-      methodNameIsDefinedByAnnotation:
-          map['methodNameIsDefinedByAnnotation'] ?? '',
+      routeNameIsDefinedByAnnotation: map['routeNameIsDefinedByAnnotation'] ?? '',
+      methodNameIsDefinedByAnnotation: map['methodNameIsDefinedByAnnotation'] ?? '',
       routeWidget: ImportableType.fromMap(map['routeWidget']),
-      returnType: map['returnType'] != null
-          ? ImportableType.fromMap(map['returnType'])
-          : null,
-      pageType: map['pageType'] != null
-          ? ImportableType.fromMap(map['pageType'])
-          : null,
+      returnType: map['returnType'] != null ? ImportableType.fromMap(map['returnType']) : null,
+      pageType: map['pageType'] != null ? ImportableType.fromMap(map['pageType']) : null,
       navigationType: NavigationType.values[map['navigationType']],
-      parameters: List<ImportableType>.from(map['parameters']?.map(
-              (dynamic x) => ImportableType.fromMap(x as Map<String, dynamic>))
-          as Iterable),
-      guards: map['guards'] == null
-          ? null
-          : List<ImportableType>.from(map['guards'].map((dynamic x) =>
-              ImportableType.fromMap(x as Map<String, dynamic>)) as Iterable),
-      defaultValues:
-          map['defaultValues'] as Map<String, dynamic>? ?? <String, dynamic>{},
+      parameters: List<ImportableType>.from(map['parameters']?.map((dynamic x) => ImportableType.fromMap(x as Map<String, dynamic>)) as Iterable),
+      guards: map['guards'] == null ? null : List<ImportableType>.from(map['guards'].map((dynamic x) => ImportableType.fromMap(x as Map<String, dynamic>)) as Iterable),
+      defaultValues: map['defaultValues'] as Map<String, dynamic>? ?? <String, dynamic>{},
+      includeQueryParameters: map['includeQueryParameters'] == null ? null : IncludeQueryParametersType.values[map['includeQueryParameters']],
     );
   }
 
   String toJson() => json.encode(toMap());
 
-  factory RouteConfig.fromJson(String source) =>
-      RouteConfig.fromMap(json.decode(source));
+  factory RouteConfig.fromJson(String source) => RouteConfig.fromMap(json.decode(source));
 }